+++ maestro-electron/vite.renderer.config.ts
@@ -13,5 +13,16 @@ export default defineConfig(async () => {
         '@': path.resolve(__dirname, './src'),
       },
     },
+    build: {
+      target: 'esnext',
+    },
+    optimizeDeps: {
+      esbuildOptions: {
+        target: 'esnext',
+        supported: {
+          'top-level-await': true
+        }
+      }
+    },
   };
 });